+++ components/ServiceType.vue
@@ -17,20 +17,22 @@
 		<section class="description" v-if="service.description">
 			<h3>Description</h3>
 			<Description :description="service.description" />
-			<DeprecationNotice v-if="service.deprecated" entity="service" />
-			<ExperimentalNotice v-if="service.experimental" entity="service" />
+			<DeprecationNotice v-if="service.deprecated" entity="type of secondary web service" />
+			<ExperimentalNotice v-if="service.experimental" entity="type of secondary web service" />
 		</section>
+
+		<FederationBackends v-if="service['federation:backends']" :backends="service['federation:backends']" :federation="federation" entity="type of secondary web service" />
 
 		<section class="configuration">
 			<h3>Configuration Settings</h3>
-			<ProcessParameter v-for="param in configuration" :key="param.name" :parameter="param" />
-			<p v-if="configuration.length === 0">This service has no configuration settings.</p>
+			<ProcessParameter v-for="param in configuration" :key="param.name" :parameter="param" :federation="federation" />
+			<p v-if="configuration.length === 0">No configuration settings available.</p>
 		</section>
 
 		<section class="parameters">
 			<h3>Process Parameters</h3>
-			<ProcessParameter v-for="param in parameters" :key="param.name" :parameter="param" />
-			<p v-if="parameters.length === 0">This service has no process parameters.</p>
+			<ProcessParameter v-for="param in parameters" :key="param.name" :parameter="param" :federation="federation" />
+			<p v-if="parameters.length === 0">No process parameters available.</p>
 		</section>
 
 		<section class="links">
@@ -44,6 +46,7 @@
 
 <script>
 import Utils from '../utils.js';
+import FederationMixin from './internal/FederationMixin.js';
 
 export default {
 	name: 'ServiceType',
@@ -54,6 +57,9 @@ export default {
 		ProcessParameter: () => import('./internal/ProcessParameter.vue'),
 		LinkList: () => import('./LinkList.vue')
 	},
+	mixins: [
+		FederationMixin
+	],
 	props: {
 		id: {
 			type: String,
@@ -62,7 +68,8 @@ export default {
 		service: {
 			type: Object,
 			default: () => ({})
-		}
+		},
+		...FederationMixin.props
 	},
 	computed: {
 		configuration() {