+++ components/ServiceTypes.vue
@@ -4,7 +4,7 @@
 			<template #heading="scope"><slot name="heading" v-bind="scope" /></template>
 			<template #summary="scope"><slot name="summary" v-bind="scope" /></template>
 			<template #details="slot">
-				<ServiceType :id="slot.summary.identifier" :service="slot.item">
+				<ServiceType :id="slot.summary.identifier" :service="slot.item" :federation="federation">
 					<template #title><span class="hidden" /></template>
 					<template #before-description="scope"><slot name="service-type-before-description" v-bind="scope" /></template>
 					<template #end="scope"><slot name="service-type-end" v-bind="scope" /></template>
@@ -16,6 +16,7 @@
 
 <script>
 import Utils from '../utils.js';
+import FederationMixin from './internal/FederationMixin.js';
 
 export default {
 	name: 'ServiceTypes',
@@ -23,6 +24,9 @@ export default {
 		SearchableList: () => Utils.loadAsyncComponent(import('./SearchableList.vue')),
 		ServiceType: () => Utils.loadAsyncComponent(import('./ServiceType.vue'))
 	},
+	mixins: [
+		FederationMixin
+	],
 	props: {
 		services: {
 			type: Object,
@@ -47,7 +51,8 @@ export default {
 		collapsed: {
 			type: Boolean,
 			default: null
-		}
+		},
+		...FederationMixin.props
 	},
 	beforeCreate() {
 		Utils.enableHtmlProps(this);