+++ components/internal/FederationMixin.js
@@ -1,0 +1,11 @@
+export default {
+	components: {
+		FederationBackends: () => import('../internal/FederationBackends.vue')
+	},
+	props: {
+		federation: {
+			type: Object,
+			default: () => ({})
+		}
+	}
+};