+++ components/internal/ProcessParameter.vue
@@ -3,12 +3,13 @@
 		<h4>
 			<code>{{ parameter.name }}</code>
 			<strong class="required" v-if="!parameter.optional" title="required">*</strong>
-			<code class="default" v-if="typeof parameter.default !== 'undefined'">&nbsp;=&nbsp;{{ defaultValue }}</code>
+			<code class="default" v-if="hasDefault">&nbsp;=&nbsp;{{ defaultValue }}</code>
 		</h4>
 		<div class="details">
 			<Description v-if="parameter.description" :description="parameter.description" :processUrl="processUrl" />
 			<DeprecationNotice v-if="parameter.deprecated === true" entity="parameter" />
 			<ExperimentalNotice v-if="parameter.experimental === true" entity="parameter" />
+			<FederationBackends v-if="parameter['federation:backends']" :backends="parameter['federation:backends']" :federation="federation" entity="parameter" />
 			<div class="json-schema-container" v-if="parameter.schema">
 				<JsonSchema :schema="parameter.schema" :processUrl="processUrl" />
 			</div>
@@ -17,19 +18,31 @@
 </template>
 
 <script>
+import FederationMixin from './FederationMixin.js';
+
 export default {
 	name: 'ProcessParameter',
-	props: {
-		parameter: Object,
-		processUrl: String
-	},
 	components: {
 		DeprecationNotice: () => import('../DeprecationNotice.vue'),
 		Description: () => import('../Description.vue'),
 		ExperimentalNotice: () => import('../ExperimentalNotice.vue'),
 		JsonSchema: () => import('../JsonSchema.vue')
 	},
+	mixins: [
+		FederationMixin
+	],
+	props: {
+		parameter: {
+			type: Object
+		},
+		processUrl: {
+			type: String
+		}
+	},
 	computed: {
+		hasDefault() {
+			return (typeof this.parameter.default !== 'undefined');
+		},
 		defaultValue() {
 			return JSON.stringify(this.parameter.default);
 		}