+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `Capabilities`: Support for the experimental Federation Extension
+- Support for the experimental Federation Extension
 
 ## [2.7.1] - 2022-01-25